[REFACTOR] Split chatroom file, rework chatroom core
