+++ pandora-client-web/src/components/chatroom/chatroom.scss
@@ -2,64 +2,34 @@
 @import '../../styles/utils';
 
 .chatroom {
-	display: flex;
-	flex-flow: row;
-	gap: 1em;
-	height: 100%;
+	overflow: hidden;
 	min-width: 0 !important;
 
-	@media only screen and (max-width: 719px) {
-		flex-flow: column;
-	}
+	>.chatroom-scene {
+		min-width: 0;
+		min-height: 0;
+		overflow: hidden;
 
-	.tab-container {
-		min-width: min(max(25em, 30vw), 100%);
-		flex: 1;
-
-		@media only screen and (max-width: 719px) {
-			flex: 6;
+		&.disabled {
+			display: none;
 		}
 	}
 
-	.controls {
-		margin: 1em;
-		flex: 1;
-	}
-}
-
-.chatroom-scene {
-	height: 100%;
-	flex: 4;
-	overflow: hidden;
-	position: relative;
-
-	&.disabled {
-		display: none;
-	}
-
-	canvas {
-		height: 100%;
-		width: 100%;
+	>.interactionArea {
+		min-width: min(25em, 100%);
 	}
 }
 
 .character-info {
+	margin-bottom: spacing('large');
+
 	.safemode {
+		display: inline-flex;
 		background: $theme-warning;
 		border: solid black 1px;
 		padding: 0.25em;
-	}
-
-	&>a {
-		padding: 0;
-	}
-
-	&>* {
-		margin-right: 1em;
-	}
-
-	&>*:last-child {
-		margin-right: 0;
+		align-items: center;
+		justify-content: center;
 	}
 }
